+++ quasim/src/ext.rs
@@ -13,14 +13,14 @@ use crate::gate::{Gate, GateType};
 macro_rules! cart {
     ($re:expr) => {
         nalgebra::Complex {
-            re: $re as f32,
+            re: $re as f64,
             im: 0.0,
         }
     };
     ($re:expr, $im:expr) => {
         nalgebra::Complex {
-            re: $re as f32,
-            im: $im as f32,
+            re: $re as f64,
+            im: $im as f64,
         }
     };
 }
@@ -28,14 +28,14 @@ macro_rules! cart {
 #[macro_export]
 macro_rules! polar {
     ($r: expr, $theta: expr) => {
-        Complex::from_polar($r as f32, $theta as f32)
+        Complex::from_polar($r as f64, $theta as f64)
     };
 }
 
 #[macro_export]
 macro_rules! cexp {
     ($exp: expr) => {
-        Complex::exp($exp as f32)
+        Complex::exp($exp as f64)
     };
 }
 
@@ -43,7 +43,7 @@ macro_rules! cexp {
 ///
 /// Two complex numbers are determined to be equal if the distance
 /// between them is at most `margin`
-pub fn equal_to_c(lhs: Complex<f32>, rhs: Complex<f32>, margin: f32) -> bool {
+pub fn equal_to_c(lhs: Complex<f64>, rhs: Complex<f64>, margin: f64) -> bool {
     (lhs - rhs).norm().le(&margin)
 }
 
@@ -57,14 +57,14 @@ pub fn equal_to_c(lhs: Complex<f32>, rhs: Complex<f32>, margin: f32) -> bool {
 /// Return `Ordering::Less` if all elements preceeding an element
 /// are equal and the same element is less
 pub fn equal_to_matrix_c<R, C, S>(
-    lhs: &Matrix<Complex<f32>, R, C, S>,
-    rhs: &Matrix<Complex<f32>, R, C, S>,
-    margin: f32,
+    lhs: &Matrix<Complex<f64>, R, C, S>,
+    rhs: &Matrix<Complex<f64>, R, C, S>,
+    margin: f64,
 ) -> bool
 where
     R: Dim,
     C: Dim,
-    S: RawStorage<Complex<f32>, R, C>,
+    S: RawStorage<Complex<f64>, R, C>,
 {
     let ((l1, l2), (r1, r2)) = (lhs.shape(), rhs.shape());
     if l1 != r1 || l2 != r2 {
@@ -88,7 +88,7 @@ pub fn reverse_indices(
 }
 
 // Helper function for get_gate_matrix
-fn u(theta: f32, phi: f32, lambda: f32) -> [Complex<f32>; 4] {
+fn u(theta: f64, phi: f64, lambda: f64) -> [Complex<f64>; 4] {
     // https://quantum.cloud.ibm.com/docs/en/api/qiskit/qiskit.circuit.library.UGate for definition
     let theta_half = theta / 2.0;
 
@@ -102,7 +102,7 @@ fn u(theta: f32, phi: f32, lambda: f32) -> [Complex<f32>; 4] {
     ]
 }
 
-pub fn get_gate2_data(gate: &Gate) -> Option<[Complex<f32>; 4]> {
+pub fn get_gate2_data(gate: &Gate) -> Option<[Complex<f64>; 4]> {
     match gate.get_type() {
         GateType::X => Some(Gate::PAULI_X_DATA),
         GateType::Y => Some(Gate::PAULI_Y_DATA),
@@ -114,16 +114,16 @@ pub fn get_gate2_data(gate: &Gate) -> Option<[Complex<f32>; 4]> {
     }
 }
 
-pub fn get_gate2_matrix(gate: &Gate) -> Option<Matrix2<Complex<f32>>> {
+pub fn get_gate2_matrix(gate: &Gate) -> Option<Matrix2<Complex<f64>>> {
     get_gate2_data(gate).map(|d| Matrix2::from_row_slice(&d))
 }
 
-pub fn get_u_matrix2(theta: f32, phi: f32, lambda: f32) -> Matrix2<Complex<f32>> {
+pub fn get_u_matrix2(theta: f64, phi: f64, lambda: f64) -> Matrix2<Complex<f64>> {
     Matrix2::from_row_slice(&u(theta, phi, lambda))
 }
 
-pub fn get_gate_matrix(gate: &Gate) -> DMatrix<Complex<f32>> {
-    let data: &[Complex<f32>] = match gate.get_type() {
+pub fn get_gate_matrix(gate: &Gate) -> DMatrix<Complex<f64>> {
+    let data: &[Complex<f64>] = match gate.get_type() {
         GateType::X => &Gate::PAULI_X_DATA,
         GateType::Y => &Gate::PAULI_Y_DATA,
         GateType::Z => &Gate::PAULI_Z_DATA,
@@ -141,7 +141,7 @@ pub fn get_gate_matrix(gate: &Gate) -> DMatrix<Complex<f32>> {
 /// Collapse a state vector into a value
 ///
 /// The sum of the squares of each item should equal to one
-pub fn collapse(state: &[Complex<f32>]) -> usize {
+pub fn collapse(state: &[Complex<f64>]) -> usize {
     let probs = state.iter().map(|&c| c.norm_sqr());
 
     let dist = WeightedIndex::new(probs)
@@ -155,16 +155,16 @@ pub fn collapse(state: &[Complex<f32>]) -> usize {
 /// Returns a probable measurement and state vector after measurement.
 pub fn measure_and_observe_sv(
     target: usize,
-    state: &DVector<Complex<f32>>,
+    state: &DVector<Complex<f64>>,
     n_qubits: usize,
-) -> (usize, DVector<Complex<f32>>) {
+) -> (usize, DVector<Complex<f64>>) {
     // Choose a collapsed state
     let prob_target_eq_zero = state
         .iter()
         .enumerate()
         .filter(|&(idx, _)| (1 << target) & idx == 0) // Using |..q_1q_0> convetion
         .map(|(_, c)| c.norm_sqr())
-        .sum::<f32>();
+        .sum::<f64>();
 
     let mut rng = rand::rng();
     let random_value = rng.random_range(0.0..1.0);
@@ -198,7 +198,7 @@ pub fn measure_and_observe_sv(
 
 /// # expand_matrix_from_gate
 /// Returns the 2^n by 2^n matrix describing a gate in a n-qubit system.
-pub fn expand_matrix_from_gate(gate: &Gate, n_qubits: usize) -> DMatrix<Complex<f32>> {
+pub fn expand_matrix_from_gate(gate: &Gate, n_qubits: usize) -> DMatrix<Complex<f64>> {
     match gate.get_type() {
         GateType::SWAP => swap_matrix(
             &gate.get_controls(),
@@ -217,15 +217,15 @@ pub fn expand_matrix_from_gate(gate: &Gate, n_qubits: usize) -> DMatrix<Complex<
 
 /// # identity_tensor_factors
 /// A `Vec` of `n_factor` number of 2 by 2 identity matricies.
-fn identity_tensor_factors(n_factors: usize) -> Vec<DMatrix<Complex<f32>>> {
-    vec![DMatrix::<Complex<f32>>::identity(2, 2); n_factors]
+fn identity_tensor_factors(n_factors: usize) -> Vec<DMatrix<Complex<f64>>> {
+    vec![DMatrix::<Complex<f64>>::identity(2, 2); n_factors]
 }
 
 /// # eval_tensor_product
 /// Evaluates the tensor product of a `Vec` of matricies.
-fn eval_tensor_product(tensor_factors: Vec<DMatrix<Complex<f32>>>) -> DMatrix<Complex<f32>> {
+fn eval_tensor_product(tensor_factors: Vec<DMatrix<Complex<f64>>>) -> DMatrix<Complex<f64>> {
     tensor_factors.iter().rev().fold(
-        DMatrix::<Complex<f32>>::identity(1, 1),
+        DMatrix::<Complex<f64>>::identity(1, 1),
         |product, factor| product.kronecker(factor),
     )
 }
@@ -237,7 +237,7 @@ pub fn swap_matrix(
     target1: usize,
     target2: usize,
     n_qubits: usize,
-) -> DMatrix<Complex<f32>> {
+) -> DMatrix<Complex<f64>> {
     /* This swap gate is implemented by a series of CNOT gates,
      *
      * 1 --*--X--*--
@@ -264,9 +264,9 @@ pub fn reverse_matrix(
     controls: &[usize],
     targets: &[usize],
     n_qubits: usize,
-) -> DMatrix<Complex<f32>> {
+) -> DMatrix<Complex<f64>> {
     let dim = 1 << n_qubits;
-    let mut mat = DMatrix::<Complex<f32>>::identity(dim, dim);
+    let mut mat = DMatrix::<Complex<f64>>::identity(dim, dim);
     for i in 0..(n_qubits >> 1) {
         mat *= swap_matrix(controls, targets[i], targets[n_qubits - 1 - i], n_qubits);
     }
@@ -276,21 +276,21 @@ pub fn reverse_matrix(
 /// # convention_convertion_matrix
 /// Returns the 2^n by 2^n matrix that converts a state vector/density matrix between,
 /// little-endian and big-endian convention. |q_0 q_1 q_2> <-> |q_2 q_1 q_0>.
-pub fn convention_convertion_matrix(n_qubits: usize) -> DMatrix<Complex<f32>> {
+pub fn convention_convertion_matrix(n_qubits: usize) -> DMatrix<Complex<f64>> {
     reverse_matrix(&[], &(0..n_qubits).collect::<Vec<usize>>(), n_qubits)
 }
 
 /// # convert_vector
 /// converts a state vector between little-endian and big-endian convention. |q_0 q_1 q_2> <-> |q_2 q_1 q_0>.
-pub fn convert_vector(vector: &DVector<Complex<f32>>) -> DVector<Complex<f32>> {
-    let n_qubits = (vector.nrows() as f32).log2() as usize;
+pub fn convert_vector(vector: &DVector<Complex<f64>>) -> DVector<Complex<f64>> {
+    let n_qubits = (vector.nrows() as f64).log2() as usize;
     convention_convertion_matrix(n_qubits) * vector
 }
 
 /// # convert_matrix
 /// converts a matrix between little-endian and big-endian convention. |q_0 q_1 q_2> <-> |q_2 q_1 q_0>.
-pub fn convert_matrix(matrix: &DMatrix<Complex<f32>>) -> DMatrix<Complex<f32>> {
-    let n_qubits = (matrix.nrows() as f32).log2() as usize;
+pub fn convert_matrix(matrix: &DMatrix<Complex<f64>>) -> DMatrix<Complex<f64>> {
+    let n_qubits = (matrix.nrows() as f64).log2() as usize;
     let mat = convention_convertion_matrix(n_qubits);
     let adj = mat.adjoint();
     mat * matrix * adj
@@ -299,11 +299,11 @@ pub fn convert_matrix(matrix: &DMatrix<Complex<f32>>) -> DMatrix<Complex<f32>> {
 /// # expand_matrix
 /// Returns the 2^n by 2^n matrix describing a gate in a n-qubit system.
 pub fn expand_matrix(
-    matrix_2x2: DMatrix<Complex<f32>>,
+    matrix_2x2: DMatrix<Complex<f64>>,
     controls: &[usize], //TODO: Allow for neg_controls.
     targets: &[usize],
     n_qubits: usize,
-) -> DMatrix<Complex<f32>> {
+) -> DMatrix<Complex<f64>> {
     let ketbra = [
         dmatrix![cart!(1.0), cart!(0.0); cart!(0.0), cart!(0.0)], // |0><0|
         dmatrix![cart!(0.0), cart!(0.0); cart!(0.0), cart!(1.0)], // |1><1|
@@ -316,7 +316,7 @@ pub fn expand_matrix(
     //   + |1><1| * |1><1| * U
     let n_terms = 1 << controls.len();
     let dim = 1 << n_qubits;
-    let mut sum = DMatrix::<Complex<f32>>::zeros(dim, dim);
+    let mut sum = DMatrix::<Complex<f64>>::zeros(dim, dim);
 
     for i in 0..n_terms {
         let mut term = identity_tensor_factors(n_qubits);
@@ -428,7 +428,7 @@ mod tests {
     };
     use crate::gate::{Gate, GateType};
     use nalgebra::{dmatrix, dvector};
-    use std::f32::consts::FRAC_1_SQRT_2;
+    use std::f64::consts::FRAC_1_SQRT_2;
 
     #[test]
     fn swap_test() {
